+++ obspy/core/event/catalog.py
@@ -34,8 +34,7 @@ from obspy.core.utcdatetime import UTCDateTime
 from obspy.core.util import NamedTemporaryFile, _read_from_plugin
 from obspy.core.util.base import (ENTRY_POINTS, download_to_file,
                                   sanitize_filename)
-from obspy.core.util.decorator import (map_example_filename, rlock,
-                                       uncompress_file)
+from obspy.core.util.decorator import map_example_filename, uncompress_file
 from obspy.core.util.misc import buffered_load_entry_point
 from obspy.imaging.cm import obspy_sequential
